services/thumbnail_generator/requirements.txt
fastapi
uvicorn
pillow  # библиотека для обработки изображений
python-multipart